requirements.txt
Flask==1.1.2
python-docx==0.8.11
textract==1.6.5
symspellpy==6.7.6
openpyxl==3.0.10
pandas==1.4.2
bitarray==2.4.1
mmh3==3.0.0
pyxdameraulevenshtein==1.7.0
mysqlclient==2.1.1
flask-mysqldb==1.0.1
